ha_requirements.txt
# HA-SentinelAI Dependencies
aiohttp>=3.9.0
pyyaml>=6.0
dnspython>=2.4.0
